The default main toolbar now contains only 5 items: Zoom in, Zoom out, Find object, Start/Stop clock, and the Timestep widget.
I have also added the ability to edit the toolbars.  However, due to a 
bug in KEditToolbar, this is currently non-functional.  You can see and 
interact with the dialog (Settings|Toolbars|Configure toolbars...), but 
your changes won't be passed to the application.

I have submitted a patch for this bug to kde-core-devel:
http://lists.kde.org/?l=kde-core-devel&m=115275435208273&w=2
You can apply the patch to your local copy if you want the edit toolbars 
tool to work now. 

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=561875