Adding the INDI observer pattern which would allow inter-driver communication when required. This feature is needed when controlling multiple devices in observatories, where drivers may watch the state and value changes of other properties they're interested in to perform their operation.
The classical example of this problem is the Rain Collector and Roof Top
(Dome)
drivers. The roof top needs to know from the rain collector driver if
it's raining or not so it can make decisions on opening or closing the
roof top. Therefore, such decisions can be made without an operator
present.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=588011