Adding LineListComponent, which is a new base class for Components in the sky which consist of collections of SkyLines (the SkyLine class was introduced a few months ago for the upcoming earth-satellite support, but it is also quite appropriate for other Components).
Many of the components that used to be PointListComponents are
now LineListComponents: EquatorComponent, EclipticComponent, and
ConstellationLineCompnent.  Using SkyLines in these components
has made the code much simpler.  There are other Components that
I plan to convert from PointListComponent to
LineListComponent (HorizonComponent, MilkyWayComponent,
CoordinateGridComponent), but the conversion is slightly more
complicated in these cases, so I haven't done it yet.

To support the new SkyLine class, I have added two functions to
SkyMap: toScreen( SkyLine*, double scale, bool useRefraction )
and checkVisibility( SkyLine* ).  These are very similar to their
SkyPoint* counterparts.  

toScreen( SkyLine* ) returns a QLineF that represents the SkyLine
transformed to pixel coordinates in the SkyMap widget.
Additionally, if the input SkyLine is only partially on-screen,
toScreen() will actually return the on-screen segment of the
SkyLine!  This is very handy, it allowed me to throw out a lot of
code in the Components that was devoted to calculating on-screen 
segments of lines.

checkVisibility( SkyLine* ) returns true if any portion of the
SkyLine is likely to be on-screen, otherwise it returns false.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=601654