CoordinatGridComponent is now a LineListComponent.
Also fixed some problems with the Milky Way, which 
were caused by a reversal of the meaning og a bool 
argument in toScreen().  It used to take a 
"clipped" argument that was set to "true" if the point 
was more than 90 degrees from the focus position (and 
therefore on the "front side" of the projected sphere 
and not to be drawn).  I decided it made more sense to
make this variable "onscreen", which is true if the 
screen point falls within the bounds of the SkyMap 
rect().  

I had forgotten to switch MilkyWayComponent to use this 
new sense of the bool variable, which was causing some 
drawing problems.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=601655