Several changes:
Changed XYZ stuff in SkyPoint and SkyMap to use Quaternions; added 
quaternion.[h|cpp] as part of this.  We still aren't actually using this 
code, but it's in place for when we are ready to try it.

Removed SatelliteTrack class; moved its functionality to 
SatelliteComponent.  Added SatelliteComposite to manage the 
SatelliteComponents.  The code is ready to start testing actual 
satellite tracks in the sky map.

MilkyWayComponent is now derived from LineListComponent.  I tried to  
make this class work properly with the new SkyMap::toScreen(SkyLine) 
function, but it got too complicated.  The problem is that this function 
truncates lines at the edge of the screen, which messes up the 
construction of the MW polygons.  Rather that try to deal with all the 
corner cases, I simply added a bool option to SkyMap::toScreen(SkyLine) 
that allows you to deactivate the truncation of lines at the screen 
edge.  So Milky Way polygons will exist beyond the borders of the 
SkyMap.  

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=607868