Fix compile.  The new cfitsio shared lib needs to be linked against libm.  Also, there's something wrong with this #ifdef wrapper in indicom.h:
#ifdef HAVE_TERMIOS_H
#include <termios.h>
#endif

I have termios.h, but it isn't being included by the above code.  This 
is on Linux (gentoo).    As a workaround, I added a duplicate #include 
to indicom.c.  This will break compilation on w32, but I'll leave it to 
the windows guys to figure out what is wrong with that ifdef.

CCMAIL: kstars-devel@kde.org
CCMAIL: Ralf.Habacker@freenet.de


svn path=/trunk/KDE/kdeedu/kstars/; revision=614947