Simplified the angular distance calc module.  Removed the "Calculate" and "Clear" buttons, and added "Find Object" buttons for the start and end positions.
Whenever a coordinate value is changed, it will check to see if four 
valid coordinate values are present.  If so, the angular distance 
between the two points will be displayed.  I use the "editingFinished()" 
signal instead of "textChanged()", so the angular value is only updated 
when the widget loses focus.

The Find Object buttons behave as you would expect.  They also add the 
selected object's name to the title of the group box enclosing the 
widgets for each position.

CCMAIL: kstars-devel@kde.org
 

svn path=/trunk/KDE/kdeedu/kstars/; revision=615134