Two fixes for the find dialog:
+ use setCurrentItem() instead of setItemSelected() when setting the 
selected item to the default ("Andromeda Galaxy").  Prior to this 
change, any interaction with the list widget caused it to jump to the 
top of the list.

+ Add "short" comet names to SkyMapComposite's ObjectNames list.  The 
short names don't include the "C/" or "P/" prefix.  This change makes it 
easier to search for comets in the find dialog.

CCMAIL:kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=624954