(1) Rename "Day Duration" calculator module to "Almanac", since it includes much more than just the length of the day
(2) Simplify UI of "Equinoxes and Solstices" module, and make it work.
You don't have to select an Equinox/Solstice anymore, the module shows 
the time/date of occurence for all four events of the given year, and 
also shows the duration of each season in days.

TODO: Right now, the seasonal names (Spring, Summer, Autumn, Winter) 
assume the northern hemisphere.  I'm considering adding a checkbox for 
this to allow users to switch to Southern hemisphere names.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=634724