New feature:  Earth Satellite tracks!
The display shows a curve across the sky, which is the 
path of a satellite over the next hour from the current 
location.  The color of the curve can be modified, and 
you can toggle whether satellite tracks are drawn in 
the Solar System tab of the Options window.

TODO:
+ Right now, only three satellites are included: HST, 
ISS, and UARS.  Many more are available, these are just 
for testing the feature.

+ Need to add satellite data to the Get New Stuff mechanism

+ Add tickmarks to the tracks, indicating the time of the 
passage

+ The path should be darkened where the satellite is in 
Earth's shadow (this information is already available, it's 
just a matter of using it).

+ Make thickness of the line proportional to brightness of 
the satellite?

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=653680