Major Change: Added Spatial Index
Structure of new classes (all in skycomponents/ except HTMesh):

  + HTMesh        (in htmesh/)
   - SkyMesh

  + LineListIndex
    - CoordinateGrid
    - ConstellationLines
    - ConsellationBoundary
    + SkipListIndex
      - MilkWay

  + PolyListIndex
    - ConstellationBoundaryPoly

Each index class has its own data container class:

  LineListIndex    SkipListIndex       PolyListIndex
       ||               ||                  ||
    LineList         SkipList            PolyList

Gory details in skycomponents/README.indexing.


Large Change: KSFileReader

I rewrote KSFileReader so it no longer tries to cache the entire
file it is reading.  I also added the ability to emit periodic
progress reports.  I consolidated the data files for stars; Deep
Sky Objects; and Constellation Boundaries; into: stars.dat, ngcic.dat,
and cbounds.dat.  Also changed the format of hipNNN.dat so the fixed
length gname comes before the variable length name.  Added conversion
program reformat-hip.pl.


New Feature: SkyLabeler (in skycomponents/)

This class is designed to prevent label overlap.  It is currently
implemented for star names; constellation names; and asteroid names.


Misc. Bug Fixes:

  o Fixed SkyMapComposite::emitProgressText()

  o Fixed: SkyMap::checkVisibility(). It was not drawing
    objects when the horizon was filled but not "visible".


Cleanup:

The new KSFileReader allowed much cleaner code in the classes that
needed to keep their data in multiple files.  I also tried to clean
up the skycomponent constructors that didn't need a visibleMethod
function pointer.  I added a "virtual bool selected()" to SkyComponent
that is now used instead in many classes.  The word "visible" is the
most overloaded word in our SkyComponent vocabulary.  "Selected" is
less ambiguous.  This also makes the code cleaner and more logical
since Options::showMilkyWay() is now in the MilkyWay class where it
belongs and not in the SkyMapComposite class, etc.

The visibleMethod mechanism still exists in some classes and may
still be needed if there are multiple instances of the same class
which need different visibleMethod functions.  If such a need exists
my plan is to make the visibleMethod mechanism specific to just the
classes that need it which will make it trivial to integrate the two
mechanisms.

The biggest change occured in DeepSkyComponent::draw().  It is
easier to understand now.  I think DSC could be further cleaned
by having a DeepSkyCatalog class with four instances but I'm in
no rush to do this now.  This is all part of a long-term effort
to remove needless complexity in skycomponents/ which should make
the code easier to understand and maintain.


-- James B. Bowlin

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=693981