NOTE: this is the HTM branch, not the trunk.
Preparing for primetime. 

The indexing needed three tweaks in order to be ready for the masses:

 1) Reverse precession correction in the SkyMesh::aperture() call.
 2) Fine tuning of the LineList JIT updating.
 3) Re-indexing of stars every 150 years.

The first one is done.  The second is almost done and I expect to have
the third one done by the end of today.

The reason the 2nd task is unfinished is because I don't know if the
constellation boundaries should precess along with the stars or if
they should be fixed to the Earth's axis like the coordinate grid.

If anyone knows if the constellation boundaries should precess or
not, please let me know and I will implement it that way.  Currently
the boundaries drawn on the screen precess but the polygons we use
to determine which constellation a point is in don't precess.
It is easy to do it either way but I can't do it both ways at the
same time (unless someone provides me with an Options::function() to
let the users choose :-).

Also, added prioritization to SkyLabeler so the most important labels
get drawn first and have a better chance of showing up since we no
longer draw labels that would overlap existing labels.  The priority
is currently:

  Planets (SolarSystemSingleComponent)
  Jupiter's Moons
  Asteroids
  Comets (which aren't being labeled ATM)
  Constellation Names
  Stars

SkyLabeler prevents labels from overlapping.  You can adjust the
vertical density with "F" and "G".  The "B" keys prints brief
statistics.  These key actions are probably for development only.
The speed and size of the SkyLabeler both scale linearly with
the number of screenRows (part of the "B" key output).


CCMAIL: kstars-devel@kde.org

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=695703