I implemented re-indexing of the stars every 150 years.  I'm haven't run extensive tests yet to verify it is doing exactly what it is supposed to but it seems to be working and hasn't broken anything. The only big surprise was that it only takes 3 to 4 seconds to re-index all the stars on a single core of an amd64 3600+.
I'm working on precessing the constellation boundary polygons but this
has turned out to be a little messy because the data is stored in
QPointF's, not SkyPoints so I need to create a new SkyPoint for every
point on the boundaries in order to do the precession.

I also need to re-index the constellation lines every 150 years just
like I do for the stars because stars form the vertices of the
constellation lines.  To do this the right way I will have to change
the data members that hold the index in LineListIndex into pointers
(instead of simply values).  I'll probably do this tomorrow.

I fixed a minor bug I had created in KStarsData where I was updating
m_updateNum more frequently than necessary.

Finally, I found a bug in the constellation lines that I don't
understand.  You can see it by zooming in on Markab, Homan, and Baham
in PEGASUS and setting the time step to one minute.  The stars will
jump away from the constellation line just a bit.  I'm a little
baffled by this because the vertices of the lines are supposed to be
the actual stars so I don't understand how they could ever not line
up.  Also, you can jump back in time 1000 years and the stars are all
attached to the lines again but if you march time forward at one
minute steps, the little jumps appear again.  I don't think this is an
Earth shattering problem, it is probably due to my lack of
understanding of the subtleties of updating StarObjects and SkyPoints.
I'll try to figure this out tomorrow.

CCMAIL: kstars-devel@kde.org


svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=695797