Some precession fixes and better JIT update syncing.
Stopped precessing:
  constellation boundaries
  constellation names

Stopped doing the reverse precession correction when we use the
index to find the constellation a point is in.

Fixed a bug in StarObject::JITupdate() where I was not updating
m_updateNumID properly.

Am now syncing the updateIDs at the start of SkyMapComposite::draw().
This means that all JIT updates in a given draw cycle will be updated
exactly the same even if updating occurs in the background in the
middle of a draw cycle.

I am now investigating Epsilon Indi over millennium time scales in
order to verify that the star re-indexing (to compensate for proper
motion) is working correctly.  These investigations led to finding
and fixing some of the precession bugs above.

CCMAIL: kstars-devel@kde.org


svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=696077