Fixed a minor bug in the code that finds which constellation a point is in.  We needed to wrap the RA of the point to negative values for some of the constellations that straddle the line RA == 0.  Added a new data member to PolyList to remember which constellations need the wrapping.  Made cosmetic changes to linelist.h and skiplist.h.
Fixed a bug in the star re-indexing code.  The new indexes I was
building were not always keeping all the stars in each trixel sorted
by brightness which was making some stars appear to disappear when
the re-indexing caused them to land in a new trixel.  This gave me
a chance to thoroughly examine the numerics in the re-indexing code
and, except for this bug of course, it all appeared solid.


CCMAIL: kstars-devel@kde.org


svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=696188