The first version of the indexing code is now feature complete.
Implemented hybrid drawing for the non-precessed guides.  This was
actually pretty easy.  I picked an arbitrary threshold of 1/4 the
total number of trixels.  When we are fully zoomed out, it takes
slightly more than half the trixels to cover the visible celestial
hemisphere.  The threshold is currently set in the SkyMesh
constructor.  Please play around with it.  I've got no idea if the
current threshold is optimal.  This opens the door for adding (at
least some of) the remaining guides to the index.

Implement re-indexing of the constellation lines.  This was a lot of
work for a very tiny effect.  I don't know if anyone would have
noticed if we had not bothered.  Re-indexing of the stars and the
CLines occurs whenever the time differs by 150 years or more from the
last time they were indexed.  We always start at J2000.

This completes the precession corrections for the index.

I encourage anyone who is interested to give it a try, see if you like
the responsiveness and even try to break it.  The README.indexing file
in the skycomponents/ directory gives an introduction and outline of
the indexing.  Things got even more complicated with the various
precession corrections so I'm working on a README.precession file.

I'd also like feedback on the labels.  Labels should no longer
overlap.  The "G" key increases the label density, the "F" key
decreases it and the "B" key prints out some simple statistics.

In addition to the README.indexing file, a lot of the code is well
documented in the header files although there is usually a few days
lag between when the code is written and when it is documented.
Sometimes longer.

CCMAIL: kstars-devel@kde.org

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=696506