Cleaned up skymesh.h and linelistindex.h including documentation of new methods.  Got rid of some cruft and unused debugging hooks. streamlined the isZoomedIn() feature.
I now determine the re-indexing update interval for the Constellation
lines at init time based on the maximum proper motion of the stars in
the constellations.  This tripled the update interval.  This was a
baby step toward implementing Jason's idea of keeping a separate list
of fast moving stars that will let us re-index the bulk of the stars
at a much, much longer interval (~4,500 years).

Internationalized my strings that users will see.

Made a small correction to SkyLabeler.  I've made the minimum x distance
between runs to be the text width of "MMMM" instead of an arbitrary number
of pixels.  Labels can be closer together than this but I won't try to
squeeze a third label in between two existing labels that are already
this close.

I made sure that I got Jason's recent skypoint update before committing.

CCMAIL: kstars-devel@kde.org


svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=696736