Implemented Jason's idea to keep some of the high proper-motions stars in a separate list and only update stars as needed.  Created a new class: HighPMStarList to do this.  If anyone has a better name for this class I'm willing to change it.  It is trivial to have multiple such lists.  Currently there are two:
  max-PM   Update Interval   # of stars
  7058      212 years             125
   838     2788 years           1,127
   304     4934 years         125,000 

The last line in the table is for the stars that are not in either
list.  If those need updating then the lists are ignored and the
entire star index is rebuilt. 

It would be trivial to add a third (or more) list, just one line of
code in the StarComponent::init() routine, but I didn't think users
would make enough multi-mellennia time jumps to justify the extra
storage space of a list with roughly 10k pointers.  If anyone has an
opinion on where we should set the proper-motion thresholds, please
let me know.  You can create a list of the stars sorted by
proper-motion by running: ./read-hippo.pl stars.data > stars.by-pm in
the data directory.  Each line in the new file will be a copy of a
line in stars.dat with the magnitude of the proper-motion prepended.


I also fixed the faint star magnitude spinner boxes in the setup menu.
The zoomed in limit was reading in new data every time the limit was
increased and the zoomed out limit was redrawing the screen every time
it was changed.  Now the stars.dat file is re-read only after the user
hits "Okay" or "Apply".  I'm actually doing the reading at the top of
StarComponent::draw() because I didn't know where the "Apply" hook
was.  But I think it is better where it is now because I can play the
same trick I did with re-indexing and delay one draw cycle before
reading the data which will allow us to present a pretty face while
the data gets read in.

Unfortunately, this still seems a little broken because it can take a
very long time to read in the star data. (I'm already using seek()
and pos() to only read in the new stars and not the entire file).  IMO
the long delay before the screen gets drawn is a big improvement over
having to battle sticky spinners but it still needs improvement.  The
4 seconds it takes to re-index all the stars is nothing compared to
the time it takes to read in the stars.dat file.

I might work on this after I port the atan2 changes. 

CCMAIL: kstars-devel@kde.org


svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=697117