I've made a correction to how we calculate proper motion in the htm branch.  I do the proper motion correction _before_ the precession correction.  We were doing it the other way around before.  The problem with the old method is that the direction of the proper motion correction would change depending on how far we had precessed.
The new method and the old method give different results.  If someone
can verify the new version is better than the old then I can port it
to the trunk.

We are still using the approximation that the proper motion is small
and is not near one of the poles but at least we are no longer
assuming that the change in coordinates due to precession is small.

CCMAIL: kstars-devel@kde.org

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=697189