The "Ecliptic" and "Equator" labels no longer overlap with each other or with any of the horizontally drawn labels.  AFAIK, the only labels that don't belong to the no-overlap club are the Horizon labels.
Now that I have the tools to prevent overlap of slanted labels it
should be pretty easy to add the Horizon labels to the club.  Most of
the code to draw slanted labels will probably be moved to SkyLabeler
when this happens to avoid code duplication.

For now, I exclude labels from the entire (vertical/horizontal)
bounding rectangle that contains a slanted label.

There is a single parameter, comfyAngle that controls the left/right
vs. low-angle priorities.  If it is zero or negative then the label
position with the minimum angle will always be used.  If it is 90 or
greater then the angle is essentially ignored.  I think the
non-overlap feature makes the left/right priority much less important.

For now, the Equator and Ecliptic labels are only being drawn when
useAntiAlias() is true.  If it is false, the labels are not drawn.

CCMAIL: kstars-devel@kde.org


svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=698075