Everything outside the Solar System is now being indexed.
The Horizon is not indexed but it is not really outside of the Solar
System and it certainly doesn't have fixed (RA, Dec).

Created LabeledListIndex (subclass of NoPrecessIndex) and created two
subclasses: Equator and Ecliptic which lets the equator and the
ecliptic make use of the HTM index.  

I still make 4 candidate label positions: left-most, right-most,
top-most, and bottom-most, but since I am making use of the SkyLabeler
to prevent overlap, I use the same priority for both the ecliptic and
equator:

  right, left, bottom, top

It would be trivial for me to give the ecliptic and the equator
different candidate priorities if that is preferred but I think it is
easier for the user if we try to consistently put those labels on the
same side of the screen. I'm totally open to changing this around if
anyone expresses a preference.  It would be easy to give the users
the option of which side to first try drawing the labels.

I reduced the number of points used to display both the ecliptic and
the equator.  The lines themselves look fine to me but when skewing,
the labels appeared to jump a little.  I then bumped up the number of
points in the equator by a factor of four and its label appears to
move much smoother.  I left the ecliptic as-is so you can compare the
jumpiness of the two.  I have no idea which way is better and am open
to suggestions.  I think the smoothly scrolling label looks better
when skewing.

The classes: EquatorComponent, EclipticComponent, and LineListComponent
are no longer being used but I haven't deleted them yet.

CCMAIL: kstars-devel@kde.org

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=698387