Once again I've made the "final" precession correction to the indexing.
The ecliptic had mistakenly become a subclass of NoPrecessIndex but the
ecliptic does precess so this had to be fixed.  The problem was in the
LabeledListIndex class which was a subclass of NoPrecessIndex.  This was
wrong because the label should have a "has a" relationship not an "is a"
relationship.  So I changed LabeledListIndex to LineListLabel and took it
out of the index hierarchy.

Both Ecliptic and Equator have a LineListLabel data member now.  Ecliptic
subclasses from LineListIndex and Equator subclasses from NoPrecessIndex.
Life is good.

The code is ugly in a couple of spots.  It will benefit greatly from
Thomas' idea of making both SkyLabeler and SkyMesh singletons.  But the
code is doing the right thing now and the structure is sound (IMO).

I may need to tweak how the ecliptic gets updated but the hooks are in
that make this trivial.

I discovered that SatelliteComponent is still subclassing from
LineListComponent so that class can't go away.  I suspect that the
satellites may never get indexed because they move so fast.

CCMAIL: kstars-devel@kde.org

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=698702