Added two routines to ConstellationBoundaryPoly and SkyMapComposite (as a proxy):
    const QPolygonF* constellationPoly( SkyPoint *p );

	const QPolygonF* constellationPoly( const QString& name );

These are the new hooks for SkyMap::HighlightConstellation() after we
merge with the trunk. 

It is clear that ConsellationBoundaryPoly should also be a singleton.
Unfortunately, there is some code in SkyMapComposite::constellationName()
that I will have to move elsewhere before I can remove that proxy method
from SkyMapComposite.

CCMAIL: kstars-devel@kde.org

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=698989