Renamed ConstellationBoundaryPoly to ConstellationBoundary and then made it into a singleton.  In order to do this I had to first rename ConstellationBoundary to ConstellationBoundaryLines.
I also had to change the cbounds.dat data file.  It used to contain
the 3-letter abbreviations for the constellation names and I replaced
those with the full names with the correct uppper/lower case.

Three files in the kstars directory had to be changed to use the
singleton.  I had already changed the call in those three files weeks
ago because I had renamed the routine so this won't make the merge
any more difficult.

I was able to get rid of proxy code in both SkyMapComposite and
in what is now ConstellationBoundaryLines which I think is a step
in the right direction.

CCMAIL: kstars-devel@kde.org

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=699040