dded code to stop drawing labels over the 3 infoBoxes.  It usually works but sometimes it doesn't especially when skewing.  I'm using the same coordinates in the InfoBoxes class that are used to draw the text strings.  It appears that those coordinates are jumping around even though the info boxes stay at fixed positions.
Fixed a couple of bugs caused by moving drawLabel() outside the draw()
routine in Ecliptic and Equator.  In general, the order we draw the
labels is the opposite of the order we draw the objects/guides
because the priority in SkyLabeler is first-come first-served which
is the opposite of the z-buffer technique we use for objects/guides.
This is why I had to move the drawLabel() routines.

Fixed a bug in SkyLabeler that would occasionally cause an index
out of bounds crash.

Fixed a bug in LineListLabeler that would cause a crash if drawLabel()
was called before reset() was ever called.

CCMAIL:kstars-devel@kde.org


svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=699181