As per Jason's suggestion, am now using the Skylabeler to draw the custom labels and the central focus label so other labels won't over overlap with them.  I've never seen the central focus label so maybe it's not working.  If it starts working again then we will need to fix the star names like the fix I implemented below for the custom labels.
These labels used to use SkyObject::drawNameLabel() which has NOT been
re-implemented with SkyLabeler.  This routine is what draws the labels
that magically appear near the cursor.  If I had used SkyLabeler here
then those labels would never appear when the normal label was already
drawn.  But since the custom labels are "attached to the skymap" they
do need to use SkyLabeler, hence the change.

I also fixed two bugs (I think) in SkyMap::drawObjectLabels().  We
were returning from the routine whenever the list contained an object
of a type that was not currently being displayed so I changed all the
return's to continue's.  Second, we weren't displaying the names of
stars properly.  For example stars with genitive names were being
labeled as "star".  I fixed this too.  Since the code to create the
label name for a star was now in three places, StarObject,
StarComponent, and SkyMap, I made StarObject::labelName so the code is
now only in one spot. I probably should have done something similar
with the offset for the star labels but I didn't.

Despite the minor bug fixes, SkyMap::drawObjectLabels() should
probably be refactored someday especially if users might make a large
number of custom labels.

CCMAIL: kstars-devel@kde.org

svn path=/branches/work/kdeedu_kstars_htm/kstars/; revision=699409