Implemented Options::hideLabels() feature for stars, asteroids, and comets, which hides the labels for these objects when slewing.
This was partly done for performance reasons because the no-overlap
feature makes drawing labels a little slower.

Currently, I'm not hiding labels for:

   Constellation Names
   Planets
   Jupiter Moons
   Ecliptic, Equator, Horizon

I don't think hiding the labels for anything on this list is required
for performance reasons.  Some of these labels, such as the planet
labels and the constellation names, are useful for orientation.

I would really like input/feedback on what labels people think should
be hidden by default when skewing since it is now an esthetics choice. 
We may need to rename the hideLabels() checkbox in the config menu to
"Most object labels".  (Sorry).

The comet labels are never displayed now partly because the AU spin
box for the label threshold is pinned to zero in the config menu.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=699595